requirements.txt
appdirs==1.4.4
asgiref==3.2.7
boto3==1.13.24
botocore==1.16.24
certifi==2020.4.5.1
chardet==3.0.4
distlib==0.3.0
dj-database-url==0.5.0
Django==3.0.7
django-crispy-forms==1.9.1
django-embed-video==1.3.2
django-filter==2.3.0
django-heroku==0.3.1
django-storages==1.9.1
django-youtube==0.2
docutils==0.15.2
filelock==3.0.12
gunicorn==20.0.4
idna==2.9
jmespath==0.10.0
mysql-connector-python==8.0.20
Pillow==7.1.2
protobuf==3.11.3
psycopg2==2.8.5
python-dateutil==2.8.1
pytz==2020.1
requests==2.23.0
s3transfer==0.3.3
six==1.15.0
sqlparse==0.3.1
urllib3==1.25.9
virtualenv==20.0.20
virtualenvwrapper-win==1.2.6
whitenoise==5.1.0
